UI Improvements
Re-factor UI functions and add more tests

Extract game setup and add tests

Extract initial-state-with-markers

Temporarily comment game setup specs to avoid infinite looping

add tests for is-a-number-on-board? and is-an-available-spot?

re-work get-user-input and get-spot

Add tests for get-marker-choice multimethod

fix confirm move to check player type

Minor UI changes to ui and game loop

Fix display winner to take a player instead of a marker